Trait graph_clustering_rs::mcl::MclExt

pub trait MclExt<A> where
    A: Float, {
    fn normalize(&self) -> Result<Array2<A>>;

    fn expand(&self, power: i32) -> Result<Array2<A>>;

    fn inflate(&self, power: A) -> Result<Array2<A>>;

    fn prune(&self, threshold: A) -> Result<Array2<A>>;

    fn add_self_loop(&mut self, loop_value: A) -> Result<()>;

    fn mcl(
        &self, 
        expansion: i32, 
        inflation: A, 
        loop_value: A, 
        iterations: usize, 
        pruning_threshold: A, 
        pruning_frequency: usize, 
        convergence_check_frequency: usize
    ) -> Result<Array2<A>>;
}

Required methods

fn normalize(&self) -> Result<Array2<A>>

Normalize the columns of the given matrix by L1 normalization

use graph_clustering_rs::mcl::*;
use ndarray::Array2;
let input: Array2<f64> = array![[1., 1., 0.],
                                [0., 1., 1.],
                                [0., 0., 1.]];
let output: Array2<f64> = array![[1., 0.5, 0.],
                                [0., 0.5, 0.5],
                                [0., 0., 0.5]];
assert_abs_diff_eq!(input.normalize().unwrap(), output)

fn expand(&self, power: i32) -> Result<Array2<A>>

Apply cluster expansion to the given matrix with given power

use graph_clustering_rs::mcl::*;
use ndarray::Array2;
let input = array![[1., 0.5, 0.],
                [0., 0.5, 0.5],
                [0., 0., 0.5]];
let output = array![[1., 0.75, 0.25],
                [0., 0.25, 0.5 ],
                [0., 0., 0.25]];
assert_abs_diff_eq!(input.expand(2).unwrap(), output) 

fn inflate(&self, power: A) -> Result<Array2<A>>

Apply cluster inflation to the given matrix with given power

use graph_clustering_rs::mcl::*;
use ndarray::Array2;
let input: Array2<f64> = array![[0.5, 0.5],
                                [1.,   1.]];
let output: Array2<f64> = array![[0.2, 0.2],
                                 [0.8, 0.8]];
assert_abs_diff_eq!(input.inflate(2.).unwrap(), output)

fn prune(&self, threshold: A) -> Result<Array2<A>>

prune the matrix below threshold The maximum value in each col is not pruned

use graph_clustering_rs::mcl::*;
use ndarray::Array2;
let threshold = 2.5;
let input: Array2<f64> = array![[1., 2., 3.], [3., 1., 4.]];
let output: Array2<f64> = array![[0., 2., 3.,], [3., 0., 4.]];
assert_abs_diff_eq!(input.prune(threshold).unwrap(), output)

fn add_self_loop(&mut self, loop_value: A) -> Result<()>

Add self loop to the matrix

fn mcl(
&self,
expansion: i32,
inflation: A,
loop_value: A,
iterations: usize,
pruning_threshold: A,
pruning_frequency: usize,
convergence_check_frequency: usize
) -> Result<Array2<A>>

mcl clustering from ndarray::Array2

use graph_clustering_rs::mcl::*;
use ndarray::Array2;

let expansion = 2;
let inflation = 2.;
let loop_value = 1.;
let iterations = 100;
let pruning_threshold = 0.0001;
let pruning_frequency = 1;
let convergence_check_frequency = 1;
let input: Array2<f64> = array![[1., 1., 1., 0., 0., 0., 0.],
                                [1., 1., 1., 0., 0., 0., 0.],
                                [1., 1., 1., 1., 0., 0., 0.],
                                [0., 0., 1., 1., 1., 0., 1.],
                                [0., 0., 0., 1., 1., 1., 1.],
                                [0., 0., 0., 0., 1., 1., 1.],
                                [0., 0., 0., 1., 1., 1., 1.]];
let output: Array2<f64> = array![[0., 0., 0., 0., 0., 0., 0.],
                                [0., 0., 0., 0., 0., 0., 0.],
                                [1., 1., 1., 0., 0., 0., 0.],
                                [0., 0., 0., 0., 0., 0., 0.],
                                [0., 0., 0., 0.5, 0.5, 0.5, 0.5],
                                [0., 0., 0., 0., 0., 0., 0.],
                                [0., 0., 0., 0.5, 0.5, 0.5, 0.5]];
assert_abs_diff_eq!(input.mcl(expansion, inflation, loop_value, iterations, pruning_threshold, pruning_frequency, convergence_check_frequency).unwrap(), output)

Loading content...